adjusted project structure
