+++ SubscriptionService/SubscriptionService.API/Extensions/ConfigurationExtenstions.cs
@@ -5,8 +5,8 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
 using SubscriptionService.API.ExceptionHandlers;
+using SubscriptionService.BusinessLogic.Features.Queries.GetAllTariffPlans;
 using SubscriptionService.BusinessLogic.Mapping;
-using SubscriptionService.BusinessLogic.Queries.GetAllTariffPlans;
 using SubscriptionService.BusinessLogic.Validators;
 using SubscriptionService.DataAccess.Data;
 using SubscriptionService.DataAccess.Repositories.Implementations;
