--- SubscriptionService/SubscriptionService.BusinessLogic/Commands/MakeSubscriptionPayment/MakeSubscriptionPaymentCommand.cs
@@ -1,15 +1,0 @@
-﻿using MediatR;
-using SubscriptionService.BusinessLogic.Models.Subscription;
-
-namespace SubscriptionService.BusinessLogic.Commands.MakeSubscriptionPayment
-{
-    public class MakeSubscriptionPaymentCommand : IRequest<GetSubscriptionDto>
-    {
-        public string Id { get; set; }
-
-        public MakeSubscriptionPaymentCommand(string id)
-        {
-            Id = id;
-        }
-    }
-}
