--- SubscriptionService/SubscriptionService.BusinessLogic/Commands/UpdateTariffPlan/UpdateTariffPlanCommand.cs
@@ -1,17 +1,0 @@
-﻿using MediatR;
-using SubscriptionService.BusinessLogic.Models.TariffPlan;
-
-namespace SubscriptionService.BusinessLogic.Commands.UpdateTariffPlan
-{
-    public class UpdateTariffPlanCommand : IRequest<GetTariffPlanDto>
-    {
-        public string Id { get; set; }
-        public UpdateTariffPlanDto Dto {  get; set; }
-
-        public UpdateTariffPlanCommand(string id, UpdateTariffPlanDto dto)
-        {
-            Id = id;
-            Dto = dto;
-        }
-    }
-}
