+++ SubscriptionService/SubscriptionService.BusinessLogic/Features/Commands/Subscription/CancelSubscription/CancelSubscriptionCommand.cs
@@ -1,0 +1,14 @@
+﻿using MediatR;
+
+namespace SubscriptionService.BusinessLogic.Features.Commands.CancelSubscription
+{
+    public class CancelSubscriptionCommand : IRequest
+    {
+        public string SubscriptionId { get; set; }
+
+        public CancelSubscriptionCommand(string subscriptionId)
+        {
+            SubscriptionId = subscriptionId;
+        }
+    }
+}
