+++ SubscriptionService/SubscriptionService.BusinessLogic/Features/Commands/Subscription/MakeSubscriptionPayment/MakeSubscriptionPaymentCommand.cs
@@ -1,0 +1,15 @@
+﻿using MediatR;
+using SubscriptionService.BusinessLogic.Models.Subscription;
+
+namespace SubscriptionService.BusinessLogic.Features.Commands.MakeSubscriptionPayment
+{
+    public class MakeSubscriptionPaymentCommand : IRequest<GetSubscriptionDto>
+    {
+        public string Id { get; set; }
+
+        public MakeSubscriptionPaymentCommand(string id)
+        {
+            Id = id;
+        }
+    }
+}
