+++ SubscriptionService/SubscriptionService.BusinessLogic/Features/Commands/Subscription/UpdateSubscription/UpdateSubscriptionCommand.cs
@@ -1,0 +1,18 @@
+﻿using MediatR;
+using SubscriptionService.BusinessLogic.Models.Subscription;
+
+namespace SubscriptionService.BusinessLogic.Features.Commands.UpdateSubscription
+{
+    public class UpdateSubscriptionCommand : IRequest<GetSubscriptionDto>
+    {
+        public string Id { get; set; }
+        public UpdateSubscriptionDto Dto { get; set; }
+
+        public UpdateSubscriptionCommand(string id, UpdateSubscriptionDto dto)
+        {
+            Id = id;
+            Dto = dto;
+        }
+
+    }
+}
