+++ SubscriptionService/SubscriptionService.BusinessLogic/Features/Queries/Subscription/GetAllSubscriptions/GetAllSubscriptionsQuery.cs
@@ -1,0 +1,16 @@
+﻿using MediatR;
+using SubscriptionService.BusinessLogic.Models;
+using SubscriptionService.BusinessLogic.Models.Subscription;
+
+namespace SubscriptionService.BusinessLogic.Features.Queries.GetAllSubscriptions
+{
+    public class GetAllSubscriptionsQuery : IRequest<PageResponse<GetSubscriptionDto>>
+    {
+        public GetPageRequest GetPageRequest { get; set; }
+
+        public GetAllSubscriptionsQuery(GetPageRequest getPageRequest)
+        {
+            GetPageRequest = getPageRequest;
+        }
+    }
+}
