+++ SubscriptionService/SubscriptionService.BusinessLogic/Features/Queries/Subscription/GetSubscriptionWithPlan/GetSubscriptionWithPlanQuery.cs
@@ -1,0 +1,18 @@
+﻿using MediatR;
+using SubscriptionService.BusinessLogic.Models;
+using SubscriptionService.BusinessLogic.Models.Subscription;
+
+namespace SubscriptionService.BusinessLogic.Features.Queries.GetSubscriptionWithPlan
+{
+    public class GetSubscriptionWithPlanQuery : IRequest<PageResponse<GetSubscriptionDto>>
+    {
+        public GetPageRequest PageRequest { get; set; }
+        public string TariffPlanName { get; set; }
+
+        public GetSubscriptionWithPlanQuery(GetPageRequest pageRequest, string tariffPlanName)
+        {
+            PageRequest = pageRequest;
+            TariffPlanName = tariffPlanName;
+        }
+    }
+}
