+++ SubscriptionService/SubscriptionService.BusinessLogic/Features/Queries/Subscription/GetUserSubscription/GetUserSubscriptionQuery.cs
@@ -1,0 +1,15 @@
+﻿using MediatR;
+using SubscriptionService.BusinessLogic.Models.Subscription;
+
+namespace SubscriptionService.BusinessLogic.Features.Queries.GetUserSubscription
+{
+    public class GetUserSubscriptionQuery : IRequest<GetSubscriptionDto>
+    {
+        public string UserName { get; set; }
+
+        public GetUserSubscriptionQuery(string userName)
+        {
+            UserName = userName;
+        }
+    }
+}
