--- SubscriptionService/SubscriptionService.BusinessLogic/Queries/GetAllSubscriptions/GetAllSubscriptionsQuery.cs
@@ -1,16 +1,0 @@
-﻿using MediatR;
-using SubscriptionService.BusinessLogic.Models;
-using SubscriptionService.BusinessLogic.Models.Subscription;
-
-namespace SubscriptionService.BusinessLogic.Queries.GetAllSubscriptions
-{
-    public class GetAllSubscriptionsQuery : IRequest<PageResponse<GetSubscriptionDto>>
-    {
-        public GetPageRequest GetPageRequest { get; set; }
-
-        public GetAllSubscriptionsQuery(GetPageRequest getPageRequest)
-        {
-            GetPageRequest = getPageRequest;
-        }
-    }
-}
