--- SubscriptionService/SubscriptionService.BusinessLogic/Queries/GetAllTariffPlans/GetAllTariffPlansQuery.cs
@@ -1,16 +1,0 @@
-﻿using MediatR;
-using SubscriptionService.BusinessLogic.Models;
-using SubscriptionService.BusinessLogic.Models.TariffPlan;
-
-namespace SubscriptionService.BusinessLogic.Queries.GetAllTariffPlans
-{
-    public class GetAllTariffPlansQuery : IRequest<PageResponse<GetTariffPlanDto>>
-    {
-        public GetPageRequest PageRequest { get; set; }
-
-        public GetAllTariffPlansQuery(GetPageRequest pageRequest)
-        {
-            PageRequest = pageRequest;
-        }
-    }
-}
