--- SubscriptionService/SubscriptionService.BusinessLogic/Queries/GetTariffPlanByName/GetTariffPlanByNameQueryHandler.cs
@@ -1,34 +1,0 @@
-﻿using AutoMapper;
-using MediatR;
-using SubscriptionService.BusinessLogic.Constants;
-using SubscriptionService.BusinessLogic.Exceptions;
-using SubscriptionService.BusinessLogic.Models.TariffPlan;
-using SubscriptionService.DataAccess.Repositories.Interfaces;
-
-namespace SubscriptionService.BusinessLogic.Queries.GetTariffPlanByName
-{
-    public class GetTariffPlanByNameQueryHandler :
-        IRequestHandler<GetTariffPlanByNameQuery, GetTariffPlanDto>
-    {
-        private readonly IUnitOfWork _unitOfWork;
-        private readonly IMapper _mapper;
-
-        public GetTariffPlanByNameQueryHandler(IUnitOfWork unitOfWork, IMapper mapper)
-        {
-            _unitOfWork = unitOfWork;
-            _mapper = mapper;
-        }
-
-        public async Task<GetTariffPlanDto> Handle(GetTariffPlanByNameQuery request, CancellationToken cancellationToken)
-        {
-            var tariffPlan = await _unitOfWork.TariffPlans.GetByNameAsync(request.Name, cancellationToken);
-
-            if (tariffPlan == null)
-            {
-                throw new NotFoundException(ExceptionMessages.TariffPlanNotFound);
-            }
-
-            return _mapper.Map<GetTariffPlanDto>(tariffPlan);
-        }
-    }
-}
