--- SubscriptionService/SubscriptionService.BusinessLogic/Queries/GetUserSubscription/GetUserSubscriptionQuery.cs
@@ -1,15 +1,0 @@
-﻿using MediatR;
-using SubscriptionService.BusinessLogic.Models.Subscription;
-
-namespace SubscriptionService.BusinessLogic.Queries.GetUserSubscription
-{
-    public class GetUserSubscriptionQuery : IRequest<GetSubscriptionDto>
-    {
-        public string UserName { get; set; }
-
-        public GetUserSubscriptionQuery(string userName)
-        {
-            UserName = userName;
-        }
-    }
-}
